+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/core/ImmutableArrayGenerator.kt
@@ -286,6 +286,7 @@ private fun generateImmutableArrayFile(baseType: BaseType): FileSpec {
                 forceFunctionBody = true,
             )
             addFilter(baseType)
+            addFilterNot(baseType)
             addPartition(baseType)
             addSortedBy(baseType)
             addSortedByDescending(baseType)
@@ -502,6 +503,30 @@ private fun TypeSpec.Builder.addFilter(baseType: BaseType) {
         }
         controlFlow("for (element in values)") {
             controlFlow("if (predicate(element))") {
+                statement("result.add(element)")
+            }
+        }
+        statement("return result.build()")
+    }
+}
+
+private fun TypeSpec.Builder.addFilterNot(baseType: BaseType) {
+    function(
+        kdoc = """
+            Returns an immutable array containing only the elements that don't match the [predicate].
+        """.trimIndent(),
+        modifiers = listOf(KModifier.INLINE),
+        name = "filterNot",
+        parameters = { "predicate"(type = lambda<Boolean> { "element"(type = baseType.type) }) },
+        returns = baseType.getGeneratedTypeName(),
+    ) {
+        if (baseType == GENERIC) {
+            statement("val result = Builder<%T>()", baseType.type)
+        } else {
+            statement("val result = Builder()")
+        }
+        controlFlow("for (element in values)") {
+            controlFlow("if (!predicate(element))") {
                 statement("result.add(element)")
             }
         }